solution to issue #8
